Set up automated tests Add gradle method to get project properties Move testing code from Robot to AutomatedTestRobot Add convenience method to wait Add AutomatedTestMain
