+++ src/main/java/frc/team2412/robot/test/AutomatedTestMain.java
@@ -1,0 +1,9 @@
+package frc.team2412.robot.test;
+
+import edu.wpi.first.wpilibj.RobotBase;
+
+public final class AutomatedTestMain {
+	public static void main(String[] args) {
+		RobotBase.startRobot(RobotTest::getInstance);
+	}
+}
